+++ CHANGELOG.md
@@ -1,9 +1,14 @@
-**202303.04**
+## 202303.05
 
+- Added in ability to scrape and include PCR results for wastewater.
+
+## 202303.04
+
+- Added in scraping of logs from the PCR table to add to wastewater submissions.
 - Completed partial imports that will add in missing reagents found in the kit indicated by the user.
 - Added web documentation to the help menu.
 
-**202303.03**
+## 202303.03
 
 - Increased robustness by utilizing PyQT6 widget names to pull data from forms instead of previously used label/input zip.
 - Above allowed for creation of more helpful prompts.